Move language attribute to app module
